+++ src/ApplicationCore/Interfaces/IRepository.cs
@@ -11,7 +11,7 @@ namespace ApplicationCore.Interfaces
 {
     public interface IRepository<T> where T:BaseEntity  
     {
-        Task<T> GetByIdAsync(int id, CancellationToken cancellationToken = default);
+        Task<T> GetByIdAsync(long id, CancellationToken cancellationToken = default);
         Task<IReadOnlyList<T>> ListAllAsync(CancellationToken cancellationToken = default);
         Task<IReadOnlyList<T>> ListAsync(ISpecification<T> spec, CancellationToken cancellationToken = default);
         Task<T> AddAsync(T entity, CancellationToken cancellationToken = default);
